requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.114.0
fastapi-cli==0.0.5
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.1
pluggy==1.5.0
psycopg==3.2.1
psycopg-binary==3.2.1
pydantic==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.2
Pygments==2.18.0
pytest==8.3.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.8.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
